refactor(app): change project struct
