+++ app/src/main/kotlin/app/theme/Color.kt
@@ -1,4 +1,4 @@
-package UIT
+package app.theme
 
 import androidx.compose.ui.graphics.Color
