+++ app/src/main/kotlin/app/ui/TreeWindow.kt
@@ -1,4 +1,4 @@
-package app
+package app.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -10,6 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import app.controller.Controller
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
